First commit in a while, read description.
I added a couple new functions, such as evaluating, compositing, together with it's enums. I added a MagickCollection class. (Nice copy from Magick.Net...) And updated the readme with instructions on how to install it on macOS and Linux.

I might expand in this some more if this project gets more traction. Thanks for reading.